requirements.txt
openai>=1.30.0
anthropic>=0.25.0
groq>=0.9.0
python-dotenv>=1.0.0
rich>=13.7.0
pydantic>=2.7.0
httpx>=0.27.0
tavily-python>=0.3.3
langsmith>=0.1.0
